requirements.txt
Django
djangorestframework
djangorestframework-simplejwt
psycopg2-binary
django-cors-headers
python-decouple
channels
channels-redis
daphne